Rename units="relative" to units="map", per section 6.7.2 of WMS 1.3.0 spec, which specifies the "Map CS", with axes "i" and "j".
Set the query popup to have scrollbars if the size > height of the map
pane.  Set the content of the popup to blank between executions of 
the query / click handler.  Bind the popup to the layer so that it 
disappears when the layer is removed from the map.

Update demo query layer mapml content title.