Add easting, northing, tcrs-x, tcrs-y, zoom as query variables.  There is some refactoring of template management needed. (Design required).
Ensure that zoomin/zoomout links are attached to the extent.  Refactoring 
required to get the processMapML business mroe DRY.

Add a hack to the leaflet.popup call to ensure that the content does not
overflow.  This required modifying mapml.css with a rule, so that will
affect the map-style.html module in web-map.